Add tests for enum types
