+++ tests/types/test_enums.py
@@ -1,70 +1,19 @@
 from enum import Enum
 
-import pytest
-
 import strawberry
-from strawberry.enum import EnumDefinition
-from strawberry.exceptions import NotAnEnum
+from strawberry.annotation import StrawberryAnnotation
 
 
-def test_raises_error_when_using_enum_with_a_not_enum_class():
-    with pytest.raises(NotAnEnum) as e:
-
-        @strawberry.enum
-        class NormalClass:
-            hello = "world"
-
-    assert ("strawberry.enum can only be used with subclasses of Enum",) == e.value.args
-
-
-def test_basic_enum():
+def test_basic():
     @strawberry.enum
-    class IceCreamFlavour(Enum):
-        VANILLA = "vanilla"
-        STRAWBERRY = "strawberry"
-        CHOCOLATE = "chocolate"
+    class NumaNuma(Enum):
+        MA = "ma"
+        I = "i"  # noqa: E741
+        A = "a"
+        HI = "hi"
 
-    definition = IceCreamFlavour._enum_definition
+    annotation = StrawberryAnnotation(NumaNuma)
+    resolved = annotation.resolve()
 
-    assert definition.name == "IceCreamFlavour"
-    assert definition.description is None
-
-    assert definition.values[0].name == "VANILLA"
-    assert definition.values[0].value == "vanilla"
-
-    assert definition.values[1].name == "STRAWBERRY"
-    assert definition.values[1].value == "strawberry"
-
-    assert definition.values[2].name == "CHOCOLATE"
-    assert definition.values[2].value == "chocolate"
-
-
-def test_can_pass_name_and_description():
-    @strawberry.enum(name="Flavour", description="example")
-    class IceCreamFlavour(Enum):
-        VANILLA = "vanilla"
-        STRAWBERRY = "strawberry"
-        CHOCOLATE = "chocolate"
-
-    definition = IceCreamFlavour._enum_definition
-
-    assert definition.name == "Flavour"
-    assert definition.description == "example"
-
-
-def test_can_use_enum_as_arguments():
-    @strawberry.enum
-    class IceCreamFlavour(Enum):
-        VANILLA = "vanilla"
-        STRAWBERRY = "strawberry"
-        CHOCOLATE = "chocolate"
-
-    @strawberry.type
-    class Query:
-        @strawberry.field
-        def flavour_available(self, flavour: IceCreamFlavour) -> bool:
-            return flavour == IceCreamFlavour.STRAWBERRY
-
-    field = Query._type_definition.fields[0]
-
-    assert isinstance(field.arguments[0].type, EnumDefinition)
+    # TODO: Remove reference to .enum_definition with StrawberryEnum
+    assert resolved is NumaNuma._enum_definition
